refac: 중복 코드 수정 및 함수 선언 통일
