+++ test-web/src/components/Stats/Charts/HeatMap/Sens_Proc_Map.js
@@ -11,27 +11,27 @@ export default function Sens_Proc_Map({
   const [chartData, setChartData] = useState({});
   const [prop, setProp] = useState([]);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await statisticSensoryProcessed(
-          startDate,
-          endDate,
-          animalType,
-          grade
-        );
+  const fetchData = async () => {
+    try {
+      const response = await statisticSensoryProcessed(
+        startDate,
+        endDate,
+        animalType,
+        grade
+      );
 
-        if (!response.ok) {
-          throw new Error('Network response was not ok');
-        }
-        const data = await response.json();
-        setProp(Object.keys(data));
-        setChartData(data);
-      } catch (error) {
-        console.error('Error fetching data:', error);
+      if (!response.ok) {
+        throw new Error('Network response was not ok');
       }
-    };
+      const data = await response.json();
+      setProp(Object.keys(data));
+      setChartData(data);
+    } catch (error) {
+      console.error('Error fetching data:', error);
+    }
+  };
 
+  useEffect(() => {
     fetchData();
   }, [startDate, endDate, animalType, grade]);
